requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp_socks==0.9.0
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
httplib2==0.22.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
oauth2client==4.1.3
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-decouple==3.8
python-socks==2.5.3
requests==2.32.3
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.13.1
